refactor(web): extract the shared enhance preset module from MediaEnhancementDrawer
Lift the preset catalog, adjustment switches, PresetCard and the params
builder (including the isCustomized rule that gates `intent: 'custom'`)
into components/media/enhancePresets.tsx so the upcoming multi-photo
batch dialog consumes the same definitions instead of a second copy.

No behavior change: buildEnhanceParams is the drawer's buildParams()
verbatim, reading a form-state object rather than closing over state.
